Extraer modelo de user a calendar
