+++ ppf/common/models/user.py
@@ -172,13 +172,3 @@ class Report(models.Model):
 
     class Meta:
         app_label = "common"
-
-
-class GoogleOAuth2Token(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    access_token = models.CharField(max_length=255)
-    refresh_token = models.CharField(max_length=255)
-    expires_at = models.DateTimeField(null=True, blank=True)
-
-    class Meta:
-        app_label = "common"
